Start by just moving these methods entirely
